moved all kind and k3s related functions into separate internal files.
